manage_documents/requirements.txt
google-cloud-firestore==2.2.0
protobuf==3.20.0
google-cloud-storage
google-cloud-logging
